Issue-3499: Centralize exception classes within each module
Partially implements #3499
